requirements.txt
debugpy
icalendar
python-dotenv
luma.led-matrix